% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listpackages.R
\name{ListFlipPackages}
\alias{ListFlipPackages}
\title{List Flip packages
This function lists the flip packages that will be loaded}
\usage{
ListFlipPackages()
}
\value{
A character vector of all known flip packages, invisibly.
}
\description{
List Flip packages
This function lists the flip packages that will be loaded
}
